+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Master
+- [New feature] Added Text type for longer than 255 strings.
 
 # 0.2.0
 - [New feature] Added kronky for better error responses on mutations