+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Master
 - [New feature] Added Text type for longer than 255 strings.
+- [New feature] Use Jason instead of Poison
 
 # 0.2.0
 - [New feature] Added kronky for better error responses on mutations