+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Master
 - [New feature] Added Text type for longer than 255 strings.
 - [New feature] Query association id in belongs_to
+- [New feature] Use Jason instead of Poison
 
 # 0.2.0
 - [New feature] Added kronky for better error responses on mutations